backend/requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.10.0
beautifulsoup4==4.13.5
bs4==0.0.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
fastapi==0.110.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
icrawler==0.6.10
idna==3.10
lxml==6.0.1
pillow==11.3.0
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
PyYAML==6.0.2
requests==2.32.5
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
starlette==0.36.3
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.29.0
watchfiles==1.1.0
websockets==15.0.1
